ON-31080 # Moved memos into a hook
